+++ Banco_Clientes_Grupal/src/main/java/org/example/rest/responses/createUpdateDelete/UserCreate.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class UserCreate {
 
     @JsonProperty("id")
-    private int id;
+    private long id;
 
     @JsonProperty("name")
     private String name;
@@ -39,7 +39,7 @@ public class UserCreate {
     @JsonProperty("company")
     private String company;
 
-    public int getId() { return id; }
+    public long getId() { return id; }
 
     public String getName() {  return name;  }
